backend/requirements.txt
Django==4.2.25
djangorestframework==3.15.2
django-cors-headers==4.4.0
pytz==2024.1
Pillow==10.4.0
python-dateutil==2.9.0

